fix: rename component and rename some of the props
